Decompose forecaster.py into models/training/evaluation modules
- Split 1,635-line services/forecaster.py: facade now 401 lines, re-exports preserved.
- New models/{config,attention,tcn,transformer,lstm}.py (each ≤300 lines).
- New training/{loaders,checkpoint,loop}.py and evaluation/metrics.py.
- Singleton state + inference path stays in services facade; train loop callbacks via lazy import.
- 426 unit + regression tests pass; ±1e-4 numerical regression locked.
