+++ backend/app/training/loaders.py
@@ -1,0 +1,265 @@
+from __future__ import annotations
+
+import csv
+import datetime
+import json
+from pathlib import Path
+from typing import Any, Sequence
+
+import torch
+
+from app.evaluation.metrics import TrainingDataSourceSummary
+from app.models.config import (
+    DEFAULT_CLOSE_SCALE,
+    DEFAULT_DATA_DIR,
+    SEQUENCE_LENGTH,
+    FeatureVector,
+)
+
+
+def _extract_required_float(record: dict[str, Any], keys: Sequence[str]) -> float:
+    for key in keys:
+        if key in record and record[key] not in {None, ""}:
+            return float(record[key])
+    raise ValueError(f"Missing required numeric field from keys: {', '.join(keys)}")
+
+
+def build_feature_vectors(
+    records: Sequence[dict[str, Any]],
+    sentiment_score: float | None = None,
+    document_date: str | None = None,
+    text_embedding: list[float] | None = None,
+) -> list[FeatureVector]:
+    vectors: list[FeatureVector] = []
+    previous_close: float | None = None
+    previous_volatility: float | None = None
+
+    parsed_doc_date: datetime.date | None = None
+    if document_date:
+        parsed_doc_date = datetime.date.fromisoformat(document_date)
+
+    sorted_records = sorted(records, key=lambda item: str(item.get("date", item.get("timestamp", ""))))
+    for record in sorted_records:
+        date_value = str(record.get("date", record.get("timestamp", "")))
+        if not date_value:
+            continue
+
+        elapsed_time = 0.0
+        if parsed_doc_date is not None:
+            record_date = datetime.date.fromisoformat(date_value[:10])
+            elapsed_time = float((record_date - parsed_doc_date).days)
+
+        close_value = _extract_required_float(record, ("close", "market_close"))
+        volatility_value = _extract_required_float(
+            record,
+            ("volatility_5d", "market_volatility", "volatility"),
+        )
+        row_sentiment = float(record.get("sentiment_score", sentiment_score if sentiment_score is not None else 0.0))
+        row_embedding = record.get("text_embedding") if isinstance(record.get("text_embedding"), list) else text_embedding
+        vectors.append(
+            FeatureVector.from_market_state(
+                date=date_value,
+                sentiment_score=row_sentiment,
+                market_close=close_value,
+                market_volatility=volatility_value,
+                previous_close=previous_close,
+                previous_volatility=previous_volatility,
+                elapsed_time=elapsed_time,
+                text_embedding=row_embedding,
+            )
+        )
+        previous_close = close_value
+        previous_volatility = volatility_value
+
+    return vectors
+
+
+def _load_json_records(path: Path) -> list[dict[str, Any]]:
+    payload = json.loads(path.read_text(encoding="utf-8"))
+    if isinstance(payload, list):
+        return [item for item in payload if isinstance(item, dict)]
+    if isinstance(payload, dict):
+        for key in ("records", "rows", "data", "items"):
+            value = payload.get(key)
+            if isinstance(value, list):
+                return [item for item in value if isinstance(item, dict)]
+    return []
+
+
+def _load_jsonl_records(path: Path) -> list[dict[str, Any]]:
+    rows: list[dict[str, Any]] = []
+    for line in path.read_text(encoding="utf-8").splitlines():
+        line = line.strip()
+        if not line:
+            continue
+        payload = json.loads(line)
+        if isinstance(payload, dict):
+            rows.append(payload)
+    return rows
+
+
+def _load_csv_records(path: Path) -> list[dict[str, Any]]:
+    with path.open("r", encoding="utf-8", newline="") as handle:
+        return [dict(row) for row in csv.DictReader(handle)]
+
+
+def _is_record_mapping_list(value: Any) -> bool:
+    return isinstance(value, list) and all(isinstance(item, dict) for item in value)
+
+
+def _extract_record_groups(payload: Any) -> list[list[dict[str, Any]]]:
+    if _is_record_mapping_list(payload):
+        if payload and any(any(key in item for key in ("records", "rows", "data", "items")) for item in payload):
+            nested_groups: list[list[dict[str, Any]]] = []
+            for item in payload:
+                if not isinstance(item, dict):
+                    continue
+                nested_groups.extend(_extract_record_groups(item))
+            return nested_groups or [payload]
+        return [payload]
+
+    if isinstance(payload, dict):
+        for key in ("sequences", "series", "groups"):
+            nested = payload.get(key)
+            if isinstance(nested, list):
+                groups: list[list[dict[str, Any]]] = []
+                for entry in nested:
+                    groups.extend(_extract_record_groups(entry))
+                if groups:
+                    return groups
+
+        for key in ("records", "rows", "data", "items"):
+            nested = payload.get(key)
+            if _is_record_mapping_list(nested):
+                return [nested]
+
+    return []
+
+
+def _load_record_groups(path: Path) -> tuple[list[list[dict[str, Any]]], str]:
+    if path.suffix == ".json":
+        payload = json.loads(path.read_text(encoding="utf-8"))
+        return _extract_record_groups(payload), "json"
+    if path.suffix == ".jsonl":
+        return [_load_jsonl_records(path)], "jsonl"
+    if path.suffix == ".csv":
+        return [_load_csv_records(path)], "csv"
+    return [], path.suffix.lstrip(".") or "unknown"
+
+
+def inspect_training_data_sources(
+    data_dir: str | Path | None = None,
+) -> tuple[list[list[FeatureVector]], list[TrainingDataSourceSummary]]:
+    root = Path(data_dir) if data_dir is not None else DEFAULT_DATA_DIR
+    if not root.exists():
+        return [], []
+
+    sequences: list[list[FeatureVector]] = []
+    summaries: list[TrainingDataSourceSummary] = []
+    for path in sorted(root.rglob("*")):
+        if not path.is_file() or path.name.startswith("."):
+            continue
+
+        try:
+            groups, format_name = _load_record_groups(path)
+            if not groups:
+                summaries.append(
+                    TrainingDataSourceSummary(
+                        path=path,
+                        format=format_name,
+                        record_groups=0,
+                        records=0,
+                        vectors=0,
+                        usable_sequences=0,
+                        status="ignored",
+                        message="No trainable market-record groups detected.",
+                    )
+                )
+                continue
+
+            record_count = sum(len(group) for group in groups)
+            vectors_for_path = [build_feature_vectors(group) for group in groups]
+            usable = [vector_group for vector_group in vectors_for_path if len(vector_group) >= SEQUENCE_LENGTH + 1]
+            sequences.extend(usable)
+            summaries.append(
+                TrainingDataSourceSummary(
+                    path=path,
+                    format=format_name,
+                    record_groups=len(groups),
+                    records=record_count,
+                    vectors=sum(len(group) for group in vectors_for_path),
+                    usable_sequences=len(usable),
+                    status="usable" if usable else "insufficient",
+                    message=(
+                        "Usable training sequences detected."
+                        if usable
+                        else f"Need at least {SEQUENCE_LENGTH + 1} usable rows per sequence."
+                    ),
+                )
+            )
+        except Exception as exc:
+            summaries.append(
+                TrainingDataSourceSummary(
+                    path=path,
+                    format=path.suffix.lstrip(".") or "unknown",
+                    record_groups=0,
+                    records=0,
+                    vectors=0,
+                    usable_sequences=0,
+                    status="error",
+                    message=str(exc),
+                )
+            )
+            continue
+
+    return sequences, summaries
+
+
+def load_training_sequences_from_data(data_dir: str | Path | None = None) -> list[list[FeatureVector]]:
+    sequences, _ = inspect_training_data_sources(data_dir)
+    return sequences
+
+
+def _build_training_tensors(
+    sequence_groups: Sequence[Sequence[FeatureVector]],
+) -> tuple[torch.Tensor | None, torch.Tensor | None]:
+    sequences: list[list[list[float]]] = []
+    targets: list[list[float]] = []
+
+    for sequence_group in sequence_groups:
+        if len(sequence_group) < SEQUENCE_LENGTH + 1:
+            continue
+        for idx in range(SEQUENCE_LENGTH, len(sequence_group)):
+            window = sequence_group[idx - SEQUENCE_LENGTH : idx]
+            target = sequence_group[idx]
+            sequences.append([item.as_list() for item in window])
+            targets.append(
+                [
+                    target.market_close / DEFAULT_CLOSE_SCALE,
+                    max(target.market_volatility, 0.0),
+                ]
+            )
+
+    if not sequences:
+        return None, None
+
+    x = torch.tensor(sequences, dtype=torch.float32)
+    y = torch.tensor(targets, dtype=torch.float32)
+    return x, y
+
+
+def _split_train_validation(
+    x: torch.Tensor,
+    y: torch.Tensor,
+    validation_split: float,
+) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
+    if len(x) < 2:
+        return x, y, x, y
+
+    val_size = max(1, int(len(x) * validation_split))
+    train_size = max(1, len(x) - val_size)
+    if train_size >= len(x):
+        train_size = len(x) - 1
+        val_size = 1
+
+    return x[:train_size], y[:train_size], x[train_size:], y[train_size:]
